aaaaaaa

+++ controles/petsController.js
@@ -2,7 +2,7 @@ const db = require('../config/db'); // Importa a conexão com o banco de dados
 
 const express = require('express');
 const router = express.Router();
-const transactionsController = require('../controles/transactionsController');
+const petsController = require('../controles/petsController');
 
 // Função para obter todas as transações 
 const getALLPets = (req, res) => { 
@@ -18,6 +18,23 @@ const getALLPets = (req, res) => {
 
 const addPet = (req, res) => {
     const{id, nome, raca, detalhes, cliente_id} = req.body;
+
+    //Verificar duplicidade
+    db.query(
+        'SELECT * FROM animal WHERE id=? AND nome=? AND raca=? AND detalhes=? AND cliente_id=?',
+        [id, nome, raca, detalhes, cliente_id],
+        (err, results) => {
+            if(err) {
+                console.error('Erro ao verificar pet', err);
+                res.status(500).send('Erro ao verificar pet');
+                return;
+            }
+            if(results.length > 0) {
+                res.status(400).send('Pet já cadastrado');
+                return;
+
+            }
+       
     db.query ('INSERT INTO animal (id, nome, raca, detalhes, cliente_id) VALUES(?, ?, ?, ?, ?)',
     [id, nome, raca, detalhes, cliente_id],
     (err,results) => {
@@ -29,7 +46,9 @@ const addPet = (req, res) => {
         res.status(201).send('Pet adicionado com sucesso');
     }
     );
-};
+}
+);
+}
 
 const updatePetPatch = (req, res) => { 
     const { id } = req.params; 
@@ -64,6 +83,13 @@ const updatePetPatch = (req, res) => {
 
        const deletePet =  (req, res) => {
         const{id} = req.params;
+
+        //Verificar se o id fornecido é válido
+        if(!id || isNaN(id)) {
+            return res.status(400).send('ID inválido ou não fornecido');
+        }
+
+        //Query para deletar o cliente
         db.query('DELETE FROM animal WHERE id = ?', [id],
             (err, results) => {
                 if(err) {
@@ -72,6 +98,7 @@ const updatePetPatch = (req, res) => {
                 return;
             }
 
+                //Verificar se o cliente foi encontrado e deletado
             if(results.affectedRows===0) {
                 res.status(404).send('Pet não encontrado');
                 return;
